IXORA-Interpretable-XAI-driven-Observational-Research-Assistant-Phase-Two/requirements.txt
# Core Dependencies
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
python-multipart>=0.0.5,<0.0.6

# Machine Learning & NLP
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
scipy>=1.7.0,<2.0.0

# Qwen Model
accelerate>=0.20.0,<0.21.0
safetensors>=0.3.0,<0.4.0
tokenizers>=0.13.0,<0.14.0
tiktoken>=0.4.0,<0.5.0

# Utilities
tqdm>=4.62.0,<5.0.0
requests>=2.26.0,<3.0.0
beautifulsoup4>=4.10.0,<5.0.0
lxml>=4.6.0,<5.0.0
pandas>=1.3.0,<2.0.0
nltk>=3.6.0,<4.0.0

# API and Web
fastapi-cors>=0.1.0,<0.2.0
httpx>=0.23.0,<0.24.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
python-multipart>=0.0.5,<0.0.6

# Development
pytest>=6.2.5,<7.0.0
pytest-asyncio>=0.15.1,<0.16.0
black>=21.12b0,<22.0.0
isort>=5.10.1,<6.0.0
mypy>=0.910,<1.0
pylint>=2.12.0,<3.0.0

# Documentation
mkdocs>=1.2.3,<2.0.0
mkdocs-material>=8.1.0,<9.0.0

# RLHF (if using reinforcement learning from human feedback)
gym>=0.21.0,<0.22.0
stable-baselines3>=1.4.0,<2.0.0

# For computer vision tasks (if needed)
opencv-python-headless>=4.5.0,<5.0.0
Pillow>=8.4.0,<9.0.0
